+++ src/pid.h
@@ -11,15 +11,15 @@ public:
 	// dt -  loop interval time
 	// max - maximum value of manipulated variable
 	// min - minimum value of manipulated variable
-	PID(double dt, int max, int min, double Kp, double Kd, double Ki);
+	PID(double dt, double max, double min, double Kp, double Kd, double Ki);
 
 	// Returns the manipulated variable given a setpoint and current process value
 	double calculate(double setpoint, double pv);
 
 private:
 	double dt;
-	int max;
-	int min;
+	double max;
+	double min;
 	double Kp;
 	double Kd;
 	double Ki;
